This branch synchronization brings many new features, bug fixes, and
optimizations. As a summary, we now have:

* Improvements in the way that `kw e` works;
* Documentation and help info enhancements;
* Organize the code installation in a Unix way;
* Dedicated deploy for Vitual Machine;
* Add support for cross-compilation;
* Introduce new features:
 - Add support for diff files;
 - DRM plugins;
 - Statistic information;

Follows the merge summary organized by authors:

Melissa Wen (3):
      src: expand git grep to cover files out git control or hidden dir
      src: diff: minor fix on help message
      src: mk: Add support for VM deploy

Rodrigo Siqueira (15):
      kw: Rename EASY_KERNEL_WORKFLOW to KWORKFLOW
      mk: Handle compilation failure
      kwlib: Add statistics track functions
      statistics: Introduce the statistics feature process
      diff: Introduce diff options
      kw_config_loader: Ignore kworkflow.config comments
      remote: Move get_remote_info() from mk to remote
      remote: Support simple bash script in the cmd_remotely
      drm: Introduce drm plugin
      setup: Rework setup
      setup: Fix kw removal from .bashrc
      setup: make installation closer to unix standards
      src: plugins: subsy: drm: Add support for load and unload driver
      documentation: man: kw.rst: Fix substitution error
      src: mk: Add support for cross-compilation

+++ src/plugins/kernel_install/arch.sh
@@ -23,36 +23,90 @@ function install_modules()
   fi
 }
 
+# Make initcpio and update grub on VM using Guestfish
+#
+# After configuring the handle (adding a disk image in write mode), the
+# guestfish performes the followed steps: (1) mount image;
+# (2) dracut (updates kernel images list); (3) create a dummy device.map
+# that tells Grub to look for /dev/sda; (4) install and update grub.
+#
+# Note: The virtual machine must be shut down and umounted before you use this
+# command, and disk images must not be edited concurrently.
+#
+# @name Kernel name for the deploy
+# @cmd_grub Command to update grub
+# mkinitcpio -g /boot/initramfs-linux.img -k /boot/vmlinuz-linux
+function vm_update_boot_loader()
+{
+  local name="$1"
+  local cmd_grub="$2"
+  local ret=0
+  local cmd=""
+  local mount_root=": mount /dev/sda1 /"
+  local cmd_init="dracut --regenerate-all -f"
+  local mkdir_grub=": mkdir-p /boot/grub"
+  local setup_grub=": write /boot/grub/device.map '(hd0,1) /dev/sda'"
+  local grub_install="grub-install --directory=/usr/lib/grub/i386-pc --target=i386-pc --boot-directory=/boot --recheck --debug /dev/sda"
+  local flag
+
+  flag=${flag:-"SILENT"}
+
+  sleep 0.5s
+
+  cmd="guestfish --rw -a ${configurations[qemu_path_image]} run \
+      $mount_root : command '$cmd_init' \
+      $mkdir_grub $setup_grub : command '$grub_install' \
+      : command '$cmd_grub'"
+
+  if [[ -f ${configurations[qemu_path_image]} ]]; then
+    warning " -> Updating grub and boot files for $name. This can take a few minutes. Please, be patient..."
+    {
+      cmd_manager "$flag" "$cmd"
+    } > /dev/null
+    say "Done."
+  else
+    complain "There is no VM in ${configurations[qemu_path_image]}"
+    return 125 # ECANCELED
+  fi
+}
+
 # Update boot loader API
 function update_boot_loader()
 {
   local name="$1"
-  local local="$2"
+  local target="$2"
   local flag="$3"
 
-  if [[ ! -z "$local" ]]; then
+  if [[ ! -z "$target" ]]; then
     sudo_cmd="sudo -E"
   fi
 
-  # Update grub
   cmd="$sudo_cmd grub-mkconfig -o /boot/grub/grub.cfg"
-  cmd_manager "$flag" "$cmd"
+
+  # Update grub
+  if [[ "$target" == 'vm' ]] ; then
+    vm_update_boot_loader "$name" "$cmd"
+  else
+    cmd_manager "$flag" "$cmd"
+  fi
 }
 
 # Install kernel
 function install_kernel()
 {
   local name="$1"
-  local reboot="$2"
-  local local="$3"
+  local kernel_image_name="$2"
+  local reboot="$3"
   local architecture="$4"
-  local flag="$5"
+  local target="$5"
+  local flag="$6"
   local sudo_cmd=""
-  local LOCAL_KW_ETC="$etc_files_path/template_mkinitcpio.preset"
+  local LOCAL_KW_ETC="$KW_ETC_DIR/template_mkinitcpio.preset"
+  local path_prefix=""
 
   flag=${flag:-"SILENT"}
 
-  if [[ "$local" == 'local' ]]; then
+  if [[ "$target" == 'local' ]]; then
     sudo_cmd="sudo -E"
   fi
 
@@ -61,38 +115,56 @@ function install_kernel()
     return 22
   fi
 
+  if [[ "$target" == 'vm' ]]; then
+    # Check if vm is mounted and get its path
+    if [[ $(findmnt "${configurations[mount_point]}") ]]; then
+      path_prefix="${configurations[mount_point]}"
+    else
+      complain "Did you check if your VM is mounted?"
+      return 125 # ECANCELED
+    fi
+  fi
+
   # Copy kernel image
-  if [[ -f "/boot/vmlinuz-$name" ]]; then
-    cmd="$sudo_cmd cp /boot/vmlinuz-$name /boot/vmlinuz-$name.old"
+  if [[ -f "$path_prefix/boot/vmlinuz-$name" ]]; then
+    cmd="$sudo_cmd cp $path_prefix/boot/vmlinuz-$name $path_prefix/boot/vmlinuz-$name.old"
     cmd_manager "$flag" "$cmd"
   fi
 
-  if [[ ! -z "$local" ]]; then
+  if [[ ! -z "$target" ]]; then
     [[ -z "$architecture" ]] && architecture="x86_64"
-    cmd="$sudo_cmd cp -v arch/$architecture/boot/bzImage /boot/vmlinuz-$name"
+    cmd="$sudo_cmd cp -v arch/$architecture/boot/$kernel_image_name /boot/vmlinuz-$name"
     cmd_manager "$flag" "$cmd"
   else
-    cmd="$sudo_cmd cp -v vmlinuz-$name /boot/vmlinuz-$name"
+    cmd="$sudo_cmd cp -v vmlinuz-$name $path_prefix/boot/vmlinuz-$name"
     cmd_manager "$flag" "$cmd"
   fi
 
   # Update mkinitcpio
-  if [[ ! -z "$local" ]]; then
-    cmd="$sudo_cmd cp -v $LOCAL_KW_ETC /etc/mkinitcpio.d/$name.preset"
+  if [[ ! -z "$target" ]]; then
+    cmd="$sudo_cmd cp -v $LOCAL_KW_ETC $path_prefix/etc/mkinitcpio.d/$name.preset"
     cmd_manager "$flag" "$cmd"
-    cmd="$sudo_cmd sed -i -e \"s/NAME/$name/g\" \"/etc/mkinitcpio.d/$name.preset\""
+    cmd="$sudo_cmd sed -i -e \"s/NAME/$name/g\" \"$path_prefix/etc/mkinitcpio.d/$name.preset\""
     cmd_manager "$flag" "$cmd"
   else
-    cp -v "$name.preset" /etc/mkinitcpio.d/
+    cp -v "$name.preset" $path_prefix/etc/mkinitcpio.d/
   fi
 
-  cmd="$sudo_cmd mkinitcpio -p $name"
-  cmd_manager "$flag" "$cmd"
+  if [[ "$target" != 'local' ]]; then
+    # Update mkinitcpio
+    cmd="$sudo_cmd mkinitcpio -p $name"
+    cmd_manager "$flag" "$cmd"
+  fi
 
-  update_boot_loader "$name" "$local" "$flag"
+  # If VM is mounted, umount before update boot loader
+  if [[ "$target" == 'vm' ]]; then
+    [[ $(findmnt "${configurations[mount_point]}") ]] && vm_umount
+  fi
+
+  update_boot_loader "$name" "$target" "$flag"
 
   # Reboot
-  if [[ "$reboot" = "1" ]]; then
+  if [[ "$target" != 'vm' && "$reboot" == "1" ]]; then
     cmd="$sudo_cmd reboot"
     cmd_manager "$flag" "$cmd"
   fi
